Moved new snapshot recording code into 'recording' package
