+++ src/main/java/com/github/tomakehurst/wiremock/junit/Stubbing.java
@@ -15,8 +15,7 @@
  */
 package com.github.tomakehurst.wiremock.junit;
 
-import com.github.tomakehurst.wiremock.admin.model.SnapshotSpec;
-import com.github.tomakehurst.wiremock.admin.model.SnapshotSpecBuilder;
+import com.github.tomakehurst.wiremock.recording.SnapshotSpecBuilder;
 import com.github.tomakehurst.wiremock.client.MappingBuilder;
 import com.github.tomakehurst.wiremock.matching.RequestPatternBuilder;
 import com.github.tomakehurst.wiremock.stubbing.ServeEvent;
