+++ src/main/java/com/github/tomakehurst/wiremock/recording/ResponseDefinitionBodyMatcherDeserializer.java
@@ -1,4 +1,4 @@
-package com.github.tomakehurst.wiremock.admin.model;
+package com.github.tomakehurst.wiremock.recording;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
