+++ src/test/java/com/github/tomakehurst/wiremock/admin/model/SnapshotRecordResultDeserialiserTest.java
@@ -1,6 +1,7 @@
 package com.github.tomakehurst.wiremock.admin.model;
 
 import com.github.tomakehurst.wiremock.common.Json;
+import com.github.tomakehurst.wiremock.recording.SnapshotRecordResult;
 import com.github.tomakehurst.wiremock.stubbing.StubMapping;
 import org.junit.Test;
