+++ src/test/java/com/github/tomakehurst/wiremock/recording/ProxiedServeEventFiltersTest.java
@@ -1,6 +1,6 @@
-package com.github.tomakehurst.wiremock.admin;
+package com.github.tomakehurst.wiremock.recording;
 
-import com.github.tomakehurst.wiremock.admin.model.ProxiedServeEventFilters;
+import com.github.tomakehurst.wiremock.recording.ProxiedServeEventFilters;
 import com.github.tomakehurst.wiremock.client.ResponseDefinitionBuilder;
 import com.github.tomakehurst.wiremock.http.ResponseDefinition;
 import com.github.tomakehurst.wiremock.matching.MockRequest;
