+++ src/test/java/com/github/tomakehurst/wiremock/recording/ResponseDefinitionBodyMatcherTest.java
@@ -1,6 +1,6 @@
-package com.github.tomakehurst.wiremock.admin;
+package com.github.tomakehurst.wiremock.recording;
 
-import com.github.tomakehurst.wiremock.admin.model.ResponseDefinitionBodyMatcher;
+import com.github.tomakehurst.wiremock.recording.ResponseDefinitionBodyMatcher;
 import com.github.tomakehurst.wiremock.client.ResponseDefinitionBuilder;
 import com.github.tomakehurst.wiremock.http.ResponseDefinition;
 import org.junit.Test;
